Arreglando carpetas para hacer el merge
