--- good-paws-web/components/Happy_family/Familia_felices.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Familia_felices = () => {
-  return (
-    <div>Familia_felices</div>
-  )
-}
