+++ good-paws-web/src/components/ui/Badge_w_Dot.jsx
