+++ good-paws-web/src/components/ui/FeaturedIcon.jsx
