+++ good-paws-web/styles/nb-globals.css
